requirements.txt
cycler==0.10.0
kiwisolver==1.2.0
numpy==1.18.4
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.14.0

matplotlib~=3.2.1
tensorflow~=2.2.0
Pillow~=7.1.2
seaborn~=0.10.1
h5py~=2.10.0
pandas~=1.0.3
scikit-learn~=0.23.0